events: move codebuild time types
